part 1 full

+++ libft/ft_strncmp.c
@@ -10,9 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <string.h>
+#include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+size_t	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned char	*p1;
 	unsigned char	*p2;
